Rename package to `com.ozonehis.it.commons`
